Clarify plugin details and automation stories
