+++ apps/app/src/components/tools/automation-overview.test.tsx
@@ -1,0 +1,26 @@
+// @vitest-environment jsdom
+
+import { render, screen } from "@testing-library/react";
+import { describe, expect, it } from "vitest";
+import { AutomationOverviewView } from "bb-plugin-automations/overview-view";
+
+describe("AutomationOverviewView", () => {
+  it("renders the production collection shell for an empty library", () => {
+    render(
+      <AutomationOverviewView
+        entries={[]}
+        error={null}
+        onRetry={() => {}}
+        onOpenDetail={() => {}}
+        onEnabledChange={async () => {}}
+        onCreateViaChat={() => {}}
+        activeMode="installed"
+        onModeChange={() => {}}
+      />,
+    );
+
+    expect(screen.getByRole("tab", { name: "Installed0" })).toBeTruthy();
+    expect(screen.getByText("No automations installed.")).toBeTruthy();
+    expect(screen.getByRole("button", { name: "New automation" })).toBeTruthy();
+  });
+});
